Allow empty or unset CUSTOM_CEPH_YUM_REPO.
Before this change, setting CUSTOM_CEPH_YUM_REPO worked fine. Not
setting it at all caused the stage generation to explode. Turns out that
the `$(call set_env_vars,$*)` stanzas in the Makefile, which uses
maint-lib/makelib.mk, sets some default variables that are carried into
the environment of programs. If we want a variable set, even if we want
it blank, it has to be set in set_env_vars.

Make a note in the toplevel Makefile where we might in future want to
set a default value. I'm not sure this is that useful, but one day
someone may thank me.